Moved remaining tasks to their own files
