+++ package.json
@@ -1,7 +1,6 @@
 {
   "name": "bulbscms",
   "version": "0.0.0",
-  "dependencies": {},
   "devDependencies": {
     "bower": "~1.3.2",
     "connect-livereload": "0.3.2",
@@ -52,6 +51,7 @@
     "karma-safari-launcher": "~0.1.1",
     "karma-sauce-launcher": "~0.2.10",
     "karma-script-launcher": "~0.1.0",
+    "load-grunt-config": "^0.17.1",
     "load-grunt-tasks": "~0.2.0",
     "protractor": "^0.22.0",
     "requirejs": "~2.1.11",
