+++ tasks/options/autoprefixer.js
@@ -1,0 +1,18 @@
+/**
+ * Add vendor prefixed styles.
+ */
+'ust strict';
+
+module.exports = {
+  options: {
+    browsers: ['last 1 version']
+  },
+  dist: {
+    files: [{
+      expand: true,
+      cwd: '.tmp/styles/',
+      src: '{,*/}*.css',
+      dest: '.tmp/styles/'
+    }]
+  }
+};
