+++ tasks/options/concat.js
@@ -1,0 +1,12 @@
+/**
+ * File concatenation.
+ */
+'use strict';
+
+var grunt = require('grunt');
+
+module.exports = {
+  options: {
+    separator: grunt.util.linefeed,
+  }
+};
