+++ tasks/options/coveralls.js
@@ -1,0 +1,13 @@
+/**
+ * Coverage reporting.
+ */
+'use strict';
+
+module.exports = {
+  options: {
+    debug: true,
+    coverage_dir: 'coverage',
+    force: true,
+    recursive: true
+  }
+};
