+++ tasks/options/shell.js
@@ -1,0 +1,13 @@
+/**
+ * Execute shell commands.
+ */
+'use strict';
+
+module.exports = {
+  bower_install: {
+    command: 'bower install -F --production'
+  },
+  bower_update: {
+    command: 'bower update -F --production'
+  }
+};
