+++ tasks/options/watch.js
@@ -3,23 +3,25 @@
  */
 'use strict';
 
+var config = require('../config');
+
 module.exports = {
   options: {
-    livereload: '<%= connect.options.livereload %>'
+    livereload: config.const.LIVERELOAD_PORT
   },
   livereload: {
     files: [
-      '<%= config.app %>/{,*/}*.html',
-      '<%= config.app %>/images/**/*.{png,jpg,jpeg,gif,webp,svg}',
-      '<%= config.app %>/scripts/**/*.js',
-      '<%= config.app %>/components/**/*.{html,js,less}',
-      '<%= config.app %>/shared/**/*.{html,js,less}',
+      config.paths.app + '/{,*/}*.html',
+      config.paths.app + '/images/**/*.{png,jpg,jpeg,gif,webp,svg}',
+      config.paths.app + '/scripts/**/*.js',
+      config.paths.app + '/components/**/*.{html,js,less}',
+      config.paths.app + '/shared/**/*.{html,js,less}',
       'test/spec/{,*/}*.js',
       'Gruntfile.js',
-      'app/styles/**/*.less',
-      '!app/styles/_mixins.less',
-      '!app/styles/_components.less',
-      '!app/styles/_shared.less'
+      config.paths.app + '/styles/**/*.less',
+      '!' + config.paths.app + '/styles/_mixins.less',
+      '!' + config.paths.app + '/styles/_components.less',
+      '!' + config.paths.app + '/styles/_shared.less'
     ],
     tasks: [
       'injector:less_components',
